requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
colorlog==6.7.0
discord.py==2.2.3
exceptiongroup==1.1.1
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
multidict==6.0.4
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-mock==3.10.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.2
yarl==1.9.2
